Refactor inner_resources
