+++ src/cradle/inner/core/monitoring.h
@@ -1,5 +1,5 @@
-#ifndef CRADLE_TYPING_CORE_MONITORING_HPP
-#define CRADLE_TYPING_CORE_MONITORING_HPP
+#ifndef CRADLE_INNER_CORE_MONITORING_H
+#define CRADLE_INNER_CORE_MONITORING_H
 
 namespace cradle {
