+++ src/cradle/inner/io/http_requests_internal.h
@@ -1,5 +1,5 @@
-#ifndef CRADLE_TYPING_IO_HTTP_REQUESTS_INTERNAL_H
-#define CRADLE_TYPING_IO_HTTP_REQUESTS_INTERNAL_H
+#ifndef CRADLE_INNER_IO_HTTP_REQUESTS_INTERNAL_H
+#define CRADLE_INNER_IO_HTTP_REQUESTS_INTERNAL_H
 
 #include <memory>
 #include <utility>
