+++ src/cradle/typing/core.h
@@ -3,11 +3,11 @@
 
 #include <cradle/inner/core/exception.h>
 #include <cradle/inner/core/id.h>
+#include <cradle/inner/core/monitoring.h>
 #include <cradle/typing/core/api_types.hpp>
 #include <cradle/typing/core/dynamic.h>
 #include <cradle/typing/core/flags.h>
 #include <cradle/typing/core/immutable.h>
-#include <cradle/typing/core/monitoring.h>
 #include <cradle/typing/core/preprocessed.h>
 #include <cradle/typing/core/type_definitions.h>
 #include <cradle/typing/core/type_info.h>
