+++ tests/external/get_context_id.cpp
@@ -2,7 +2,7 @@
 
 #include "test_session.h"
 #include <cradle/external_api.h>
-#include <cradle/typing/io/mock_http.h>
+#include <cradle/inner/io/mock_http.h>
 #include <cradle/typing/service/core.h>
 #include <cradle/typing/utilities/testing.h>
