+++ tests/inner/io/mock_http.cpp
@@ -1,6 +1,7 @@
-#include <cradle/typing/io/mock_http.h>
+#include <catch2/catch.hpp>
 
-#include <cradle/typing/utilities/testing.h>
+#include <cradle/inner/core/monitoring.h>
+#include <cradle/inner/io/mock_http.h>
 
 using namespace cradle;
