+++ tests/unit/thinknode/iam.cpp
@@ -2,8 +2,8 @@
 
 #include <cppcoro/sync_wait.hpp>
 
-#include <cradle/typing/core/monitoring.h>
-#include <cradle/typing/io/mock_http.h>
+#include <cradle/inner/core/monitoring.h>
+#include <cradle/inner/io/mock_http.h>
 #include <cradle/typing/service/core.h>
 #include <cradle/typing/utilities/testing.h>
