+++ tests/unit/thinknode/iss.cpp
@@ -6,9 +6,10 @@
 
 #include <cppcoro/sync_wait.hpp>
 
-#include <cradle/typing/core/monitoring.h>
+#include <cradle/inner/core/monitoring.h>
+#include <cradle/inner/io/http_requests.h>
+#include <cradle/inner/io/mock_http.h>
 #include <cradle/typing/encodings/msgpack.h>
-#include <cradle/typing/io/mock_http.h>
 #include <cradle/typing/service/core.h>
 #include <cradle/typing/utilities/testing.h>
